requirements.txt
django
mysqlclient
mysql-connector-python
djangorestframework
djangorestframework-simplejwt
